create ligand list component
