+++ baby-gru/src/components/BabyGruMoleculeCard.js
@@ -1,9 +1,10 @@
 import React, { useEffect, useState } from "react";
-import { Card, Form, Row, Col, Button } from "react-bootstrap";
+import { Card, Form, Row, Col } from "react-bootstrap";
 import { doDownload, sequenceIsValid } from '../utils/BabyGruUtils';
 import { isDarkBackground } from '../WebGL/mgWebGL'
 import { BabyGruSequenceViewer } from "./BabyGruSequenceViewer";
 import { BabyGruMoleculeCardButtonBar } from "./BabyGruMoleculeCardButtonBar"
+import { BabyGruLigandList } from "./BabyGruLigandList"
 
 export const BabyGruMoleculeCard = (props) => {
     const [showState, setShowState] = useState({})
@@ -14,8 +15,6 @@ export const BabyGruMoleculeCard = (props) => {
     const [bondWidth, setBondWidth] = useState(0.1)
     const [atomRadiusBondRatio, setAtomRadiusBondRatio] = useState(1.5)
     const [bondSmoothness, setBondSmoothness] = useState(1)
-    const [ligandList, setLigandList] = useState([])
-    const [cachedAtoms, setCachedAtoms] = useState(null)
 
     const  bondSettingsProps = {
         bondWidth, setBondWidth, atomRadiusBondRatio, 
@@ -183,38 +182,6 @@ export const BabyGruMoleculeCard = (props) => {
         props.setCurrentDropdownMolNo(-1)
     }
 
-    useEffect(() => {
-        async function updateMoleculeAtoms() {
-            await props.molecule.updateAtoms()
-        }
-
-        if (props.molecule.gemmiStructure === null || props.molecule.atomsDirty) {
-            updateMoleculeAtoms()
-        }
-        if (props.molecule.gemmiStructure === null) {
-            return
-        }
-        
-        let ligandList = []
-        const model = props.molecule.gemmiStructure.first_model()
-        const chains = model.chains
-        for (let i = 0; i < chains.size(); i++) {
-            let chain = chains.get(i)
-            const ligands = chain.get_ligands_const()
-            for (let j = 0; j < ligands.size(); j++) {
-                let ligand = ligands.at(j)
-                ligandList.push({res: ligand, chainName: chain.name})
-            }
-        }
-        setLigandList(ligandList)
-    
-    }, [cachedAtoms])
-
-    useEffect(() => {
-        setCachedAtoms(props.molecule.cachedAtoms)
-    })
-
-
     const handleProps = { handleCentering, handleCopyFragment, handleDownload, handleRedo, handleUndo, handleResidueRangeRefinement, handleVisibility}
 
     return <Card className="px-0" style={{ marginBottom: '0.5rem', padding: '0' }} key={props.molecule.molNo}>
@@ -333,34 +300,7 @@ export const BabyGruMoleculeCard = (props) => {
                     }
                 </Col>
             </Row>
-            {ligandList.length > 0 && 
-                <>
-                    <hr></hr>
-                    <Row style={{ height: '100%' }}>
-                        <Col>
-                            <div>
-                                <b>Ligands</b>
-                            </div>
-                            {ligandList.map(ligand => {
-                                return <Card style={{margin: '0.5rem'}}>
-                                            <Card.Body>
-                                                <Row style={{display:'flex', justifyContent:'between'}}>
-                                                    <Col style={{alignItems:'center', justifyContent:'left', display:'flex'}}>
-                                                        {`${ligand.chainName}/${ligand.res.seqid.str()}(${ligand.res.name})`}
-                                                    </Col>
-                                                    <Col className='col-3' style={{justifyContent: 'right', display:'flex'}}>
-                                                        <Button onClick={() => {props.molecule.centreOn(props.glRef, {chain: ligand.chainName, molName: props.molecule.name, molNo: props.molecule.molNo, modelIndex: 0, seqNum: ligand.res.seqid.str()})}}>
-                                                            View
-                                                        </Button>
-                                                    </Col>
-                                                </Row>
-                                            </Card.Body>
-                                        </Card>
-                            })}
-                        </Col>
-                    </Row>
-                </>
-            }
+            <BabyGruLigandList molecule={props.molecule} glRef={props.glRef}/>
         </Card.Body>
     </Card >
 }
